refactor(openapi): promote latency + request_fingerprint + custom_fixture — Phase 4a
Moves three infrastructure modules out of mockforge-core:

- \`latency.rs\` -> mockforge-foundation (only used \`crate::Result\`; pure
  data + distribution sampling). Core retains a re-export shim.
- \`request_fingerprint.rs\` -> mockforge-openapi (its non-external deps
  are \`OpenApiSpec\` + axum types). Naturally an OpenAPI-layer concept.
- \`custom_fixture.rs\` -> mockforge-openapi (depends on RequestFingerprint
  so follows it). Fixture matching is OpenAPI-path-aware.

These three are heavy-traffic deps of \`openapi_routes\` and
\`mockforge-http\`, so moving them clears a significant chunk of the core
surface area. The remaining blockers for moving \`openapi_routes\` itself
(templating, overrides, conditions, time_travel, encryption) are all
deep-cascade core modules that still need either promotion or hook-trait
design — out of scope for this commit.

mockforge-openapi gains an \`axum\` runtime dep (needed by RequestFingerprint
for Method/Uri/HeaderMap) and \`tempfile\` as a dev-dep (used by
custom_fixture tests). Modernises three \`rand::thread_rng()\` deprecations
in the moved latency.rs so foundation clippy-passes.

Core keeps every legacy path resolving via re-exports:
- \`mockforge_core::latency\` -> mockforge_foundation::latency
- \`mockforge_core::custom_fixture\` -> mockforge_openapi::custom_fixture
- \`mockforge_core::request_fingerprint\` -> mockforge_openapi::request_fingerprint

Verification:
- cargo build --workspace --exclude mockforge-desktop clean
- cargo test -p mockforge-foundation --lib — 114 passed (up from 84)
- cargo test -p mockforge-openapi --lib — 106 passed (up from 87)
- cargo test -p mockforge-core --lib — 1640 passed (down from 1689; 49 tests migrated with modules)
- cargo clippy --all-targets --all-features --workspace --exclude mockforge-desktop -- -D warnings clean
- cargo fmt --all --check clean

Plan: ~/.claude/plans/openapi-spec-extraction.md (Phase 4a; full
openapi_routes move still blocked)
